+++ lib/generators/ember/component_generator.rb
@@ -9,14 +9,12 @@ module Ember
 
       desc "Creates a new Ember.js component and component template\nCustom Ember Components require at least two descriptive names separated by a dash. Use CamelCase or dash-case to name your component.\n\nExample,\n\trails generate ember:component PostChart [options]\n\trails generate ember:component post-chart [options]"
 
-      class_option :ember_path, :type => :string, :aliases => "-d", :default => false, :desc => "Custom ember app path"
-
       def create_component_files
         dashed_file_name = file_name.gsub(/_/, '-')
-        component_path = File.join(ember_path, 'components', class_path, "#{dashed_file_name}_component.js.es6")
+        component_path = File.join(app_path, 'components', class_path, "#{dashed_file_name}_component.js.es6")
         template "component.js.es6", component_path
 
-        template_path = File.join(ember_path, 'templates/components', class_path, "#{dashed_file_name}.hbs")
+        template_path = File.join(app_path, 'templates/components', class_path, "#{dashed_file_name}.hbs")
         template "component.template.hbs", template_path
       end
     end