+++ test/generators/route_generator_test.rb
@@ -12,13 +12,6 @@ class RouteGeneratorTest < Rails::Generators::TestCase
   test "Assert files are properly created" do
     run_generator %w(index)
 
-    assert_file "#{ember_path}/routes/index.js.es6"
-  end
-
-  test "Assert files are properly created with custom path" do
-    custom_path = ember_path("custom")
-    run_generator [ "index", "-d", custom_path ]
-
-    assert_file "#{custom_path}/routes/index.js.es6"
+    assert_file "#{app_path}/routes/index.js.es6"
   end
 end