+++ test/generators/view_generator_test.rb
@@ -11,42 +11,36 @@ class ViewGeneratorTest < Rails::Generators::TestCase
 
   test "create view with template by default" do
     run_generator ["post"]
-    assert_file "#{ember_path}/views/post.js.es6"
-    assert_file "#{ember_path}/templates/post.hbs"
+    assert_file "#{app_path}/views/post.js.es6"
+    assert_file "#{app_path}/templates/post.hbs"
   end
 
   test "create view without a template" do
     run_generator ["post", "--without-template"]
-    assert_file "#{ember_path}/views/post.js.es6"
-    assert_no_file "#{ember_path}/templates/post.hbs"
+    assert_file "#{app_path}/views/post.js.es6"
+    assert_no_file "#{app_path}/templates/post.hbs"
   end
 
   test "create view and template (using ember-rails flags)" do
     run_generator ["post", "--with-template"]
-    assert_file "#{ember_path}/views/post.js.es6"
-    assert_file "#{ember_path}/templates/post.hbs"
+    assert_file "#{app_path}/views/post.js.es6"
+    assert_file "#{app_path}/templates/post.hbs"
   end
 
   test "create namespaced view" do
     run_generator ["post/index"]
-    assert_file "#{ember_path}/views/post/index.js.es6"
+    assert_file "#{app_path}/views/post/index.js.es6"
   end
 
   test "Assert files are properly created" do
     run_generator %w(ember)
-    assert_file "#{ember_path}/views/ember.js.es6"
+    assert_file "#{app_path}/views/ember.js.es6"
   end
 
-  test "Assert files are properly created with custom path" do
-    custom_path = ember_path("custom")
-    run_generator [ "ember", "-d", custom_path ]
-    assert_file "#{custom_path}/views/ember.js.es6"
-  end
+  test "Uses config.ember.appkit.paths.app" do
+    custom_path = app_path("custom")
 
-  test "Uses config.ember.ember_path" do
-    custom_path = ember_path("custom")
-
-    with_config ember_path: custom_path do
+    with_config paths: {app: custom_path} do
       run_generator [ "ember"]
       assert_file "#{custom_path}/views/ember.js.es6"
     end