+++ CHANGELOG.md
@@ -12,3 +12,4 @@
 * Added utils/csrf.js to autoset csrf token in $.ajaxPrefilter - Brian Cardarella
 * Added cache buster for in-gem application.hbs for Sprockets - Brian Cardarella & Lin Reid
 * Added Scaffold Generator - Brian Cardarella
+* Application generates into `app/` and `config/` - Brian Cardarella